requirements.txt
pandas
numpy
sentence-transformers
scikit-learn
torch
language-tool-python
python-dotenv
fastapi
uvicorn
scikit-learn
pydantic
tiktoken
mysql-connector-python
datasets
accelerate>=1.1.0